11-2-1-完善群组聊天的所有逻辑-完善部分界面

+++ App/ITalker/factory/src/main/java/net/qiujuer/italker/factory/presenter/message/ChatContract.java
@@ -46,6 +46,6 @@ public interface ChatContract {
         void showAdminOption(boolean isAdmin);
 
         // 初始化成员信息
-        void onInitGroupMembers(List<MemberUserModel> members, int moreCount);
+        void onInitGroupMembers(List<MemberUserModel> members, long moreCount);
     }
 }
